refactor: move health section contracts to engine
