+++ crates/api/src/lib.rs
@@ -252,11 +252,9 @@ pub struct ComplexityOptions {
     pub coverage_root: Option<PathBuf>,
 }
 
+pub use fallow_engine::{DerivedHealthSections, HealthSectionOptions, derive_health_sections};
+
 /// Derived section selection for programmatic health / complexity runs.
-///
-/// Owned by `fallow-api` so embedders and the remaining CLI-backed runtime path
-/// share the same defaulting rules while health execution moves behind the
-/// typed API boundary.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct DerivedComplexityOptions {
     pub any_section: bool,
@@ -269,91 +267,6 @@ pub struct DerivedComplexityOptions {
     pub force_full: bool,
     pub score_only_output: bool,
     pub score: bool,
-}
-
-/// Input for deriving effective health sections from command-neutral flags.
-#[derive(Debug, Clone)]
-pub struct HealthSectionOptions {
-    pub output: fallow_config::OutputFormat,
-    pub complexity: bool,
-    pub file_scores: bool,
-    pub coverage_gaps: bool,
-    pub hotspots: bool,
-    pub targets: bool,
-    pub css: bool,
-    pub score: bool,
-    pub score_gate: bool,
-    pub snapshot_requested: bool,
-    pub trend: bool,
-}
-
-/// Derived section selection for health runs.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct DerivedHealthSections {
-    pub any_section: bool,
-    pub complexity: bool,
-    pub file_scores: bool,
-    pub coverage_gaps: bool,
-    pub hotspots: bool,
-    pub targets: bool,
-    pub css: bool,
-    pub score: bool,
-    pub force_full: bool,
-    pub score_only_output: bool,
-}
-
-/// Derive effective health section flags for CLI and embedders.
-#[must_use]
-pub fn derive_health_sections(options: &HealthSectionOptions) -> DerivedHealthSections {
-    let score = options.score
-        || options.score_gate
-        || options.trend
-        || matches!(options.output, fallow_config::OutputFormat::Badge);
-    let any_section = options.complexity
-        || options.file_scores
-        || options.coverage_gaps
-        || options.hotspots
-        || options.targets
-        || score;
-    let effective_score = if any_section { score } else { true } || options.snapshot_requested;
-    let force_full = options.snapshot_requested || effective_score;
-
-    DerivedHealthSections {
-        any_section,
-        complexity: if any_section {
-            options.complexity
-        } else {
-            true
-        },
-        file_scores: if any_section {
-            options.file_scores
-        } else {
-            true
-        } || force_full,
-        coverage_gaps: if any_section {
-            options.coverage_gaps
-        } else {
-            false
-        },
-        hotspots: if any_section { options.hotspots } else { true }
-            || options.snapshot_requested
-            || options.trend,
-        targets: if any_section { options.targets } else { true },
-        css: options.css,
-        score: effective_score,
-        force_full,
-        score_only_output: is_health_score_only_output(options, score),
-    }
-}
-
-fn is_health_score_only_output(options: &HealthSectionOptions, score: bool) -> bool {
-    score
-        && !options.complexity
-        && !options.file_scores
-        && !options.coverage_gaps
-        && !options.hotspots
-        && !options.targets
-        && !options.trend
 }
 
 /// Derive effective programmatic health / complexity section flags.
